refactor(output): extract classification tests to sibling file
